Error ojs version 3.2.1.4: OptimetaCitationsPlugin and OptimetaCitationsPluginBase merged
